requirements.txt
streamlit==1.27.0
openai>=1.5.0
openrouter==0.0.19
google-api-python-client
google-auth
requests
